Migrate data fixtures into conftest for easy reuse
